+++ code/data_structures/linked_list/linked_list.go
