+++ code/data_structures/linked_list/linked_lists_operations/reverse_list.cpp
