+++ code/data_structures/linked_list/linked_lists_operations/reverselinkedlist_iteration.cpp
