+++ code/data_structures/linked_list/linked_lists_operations/rotate_list.cpp
