+++ code/data_structures/segment_tree/segment_tree_rmq.py
